834108: Set the default connection timeout to 1 min.
httplib in python <= 2.4 seems to have a very long timeout
of around 3 mins. When an invalid proxy or server address
is configured, this feels way to long in both the CLI and
UI.

This value can be changed by adding a 'connection_timeout'
property under the 'server' section of the config file.