Add *Async methods to uep API.
Add bindAsync(), bindByEntitlementPoolAsync(), and
bindByProductAsync().

Since these result in completely different return values,
split the regular and "async" varieties into two seperate
methods.

'202' status is no longer considered an error.

Request made to the the candlepin api with 'async=true'
return a 202 and a candlepin JobStatus. ValidateResponse
was previously considering 202 a failure (as it fell into
"not a 200 or 204").

Add updateJobStatus

Feed the result of a job queueing request (ie, async)
to updateStatus to check the latest status of the job.

updateStatus pulls the job href out of the status and
makes a request to it and returns the results.