Move ent cert auth out of connection into auth.
EntitlementCertRestlib was settng entitlement cert
base client cert auth in it's set, but this moves it
to a AuthBase/RhsmAuth class.

Cleaned up the ent cert searcing a bit. This is kind
of duplicating subman cert directory code though, so
it may make sense to ditch this and make that
submans release.py responsibility.