Open temp files for tests in text mode.
More specifically, the NamedTemporaryFile's we
use for the config tests. The test open the file
and write config data to it, but the config data
is currently plain py2 style strings, ie, 'bytes'
on py3 and needs an encoding for 'binary' mode files.

Better fix is likely to either specify the encoding
or mark up those strings as bytes. But opening the
file in 't' mode works for now.

See: http://stackoverflow.com/a/24860337/781180