1435475: Support older versions of M2Crypto
Older versions of M2Crypto do not support the "family" parameter in
M2Crypto.SSL.Connection.  This parameter allows one to create an IPv4 or
IPv6 connection.  This patch checks to see if "family" is supported and
if so, it provides the correct value.  Otherwise, nothing is supplied.
Usage of an older version of M2Crypto will limit python-rhsm to IPv4
but this was always the case in actual fact.  Usage of newer versions
will be able to use either IPv6 or IPv4.